+++ src/shapes/Line.ts
@@ -4,7 +4,7 @@ import type { TClassProperties } from '../typedefs';
 import { classRegistry } from '../ClassRegistry';
 import { FabricObject, cacheProperties } from './Object/FabricObject';
 import { Point } from '../Point';
-import { isFiller } from '../util/types';
+import { isFiller } from '../util/typeAssertions';
 import type {
   FabricObjectProps,
   SerializedObjectProps,
